+++ Common/ZDB/Services/UserService.cs
@@ -10,31 +10,38 @@ namespace ZetaDashboard.Common.ZDB.Services
     {
         public class UserService: MongoRepositoryBase<UserModel>
         {
+            private List<string> thispage = new List<string>() { "zdb" };
             public UserService(MongoContext context)
                 : base(context, "users") { }
 
 
             #region Get
-            public async Task<ApiResponse<List<UserModel>>> GetAllUsersAsync()
+            public async Task<ApiResponse<List<UserModel>>> GetAllUsersAsync(UserModel loggeduser)
             {
                 ApiResponse<List<UserModel?>> response = new ApiResponse<List<UserModel?>>();
                 try
                 {
+                    if (!HasPermissions(loggeduser, UserModel.EUserPermissionType.Visor, thispage))
+                    {
+                        response.Result = ResponseStatus.Unauthorized;
+                        response.Message = "No tienes permisos";
+                        return response;
+                    }
                     var result = await FindAllAsync();
                     if (result != null)
                     {
-                        response.Result = true;
+                        response.Result = ResponseStatus.Ok;
                         response.Data = result;
                     }
                     else
                     {
-                        response.Result = false;
+                        response.Result = ResponseStatus.NotFound;
                         response.Message = "Error obteniendo usuarios";
                     }
                 }
                 catch (Exception ex)
                 {
-                    response.Result = false;
+                    response.Result = ResponseStatus.InternalError;
                     response.Message = "Error" + ex.Message;
                 }
                 return response;
@@ -56,36 +63,42 @@ namespace ZetaDashboard.Common.ZDB.Services
 
                     if (first != null && BCrypt.Net.BCrypt.Verify(user.PasswordHash, first.PasswordHash))
                     {
-                        response.Result = true;
+                        response.Result = ResponseStatus.Ok;
                         response.Data = first;
                     }
                     else
                     {
-                        response.Result = false;
+                        response.Result = ResponseStatus.NotFound;
                         response.Message = "Email o contraseña incorrectos";
                     }
                 }
                 catch (Exception ex)
                 {
-                    response.Result = false;
+                    response.Result = ResponseStatus.InternalError;
                     response.Message += "Error:" + ex.Message;
                 }
 
                 return response;
             }
-            public async Task<ApiResponse<bool>> InsertUserAsync(UserModel model)
+            public async Task<ApiResponse<bool>> InsertUserAsync(UserModel model, UserModel loggeduser)
             {
                 ApiResponse<bool> response = new ApiResponse<bool>();
 
                 try
                 {
+                    if (!HasPermissions(loggeduser, UserModel.EUserPermissionType.Editor, thispage))
+                    {
+                        response.Result = ResponseStatus.Unauthorized;
+                        response.Message = "No tienes permisos";
+                        return response;
+                    }
                     await InsertAsync(model);
-                    response.Result = true;
+                    response.Result = ResponseStatus.Ok;
                     response.Message = "El usuario se ha insertado correctamente";
                 }
                 catch (Exception ex)
                 {
-                    response.Result = false;
+                    response.Result = ResponseStatus.InternalError;
                     response.Message = "Ha ocurrido un error al insertar el usuario";
                     Console.WriteLine($"Error: {ex.Message}");
                 }
@@ -94,19 +107,25 @@ namespace ZetaDashboard.Common.ZDB.Services
             #endregion
 
             #region Update
-            public async Task<ApiResponse<bool>> UpdateUserAsync(UserModel model)
+            public async Task<ApiResponse<bool>> UpdateUserAsync(UserModel model, UserModel loggeduser)
             {
                 ApiResponse<bool> response = new ApiResponse<bool>();
 
                 try
                 {
+                    if (!HasPermissions(loggeduser, UserModel.EUserPermissionType.Editor, thispage))
+                    {
+                        response.Result = ResponseStatus.Unauthorized;
+                        response.Message = "No tienes permisos";
+                        return response;
+                    }
                     await UpdateAsync(model);
-                    response.Result = true;
+                    response.Result = ResponseStatus.Ok;
                     response.Message = "El usuario se ha editado correctamente";
                 }
                 catch (Exception ex)
                 {
-                    response.Result = false;
+                    response.Result = ResponseStatus.InternalError;
                     response.Message = "Ha ocurrido un error al editar el usuario";
                     Console.WriteLine($"Error: {ex.Message}");
                 }
@@ -115,19 +134,25 @@ namespace ZetaDashboard.Common.ZDB.Services
             #endregion
 
             #region Delete
-            public async Task<ApiResponse<bool>> DeleteUserAsync(UserModel model)
+            public async Task<ApiResponse<bool>> DeleteUserAsync(UserModel model, UserModel loggeduser)
             {
                 ApiResponse<bool> response = new ApiResponse<bool>();
 
                 try
                 {
+                    if (!HasPermissions(loggeduser, UserModel.EUserPermissionType.Admin, thispage))
+                    {
+                        response.Result = ResponseStatus.Unauthorized;
+                        response.Message = "No tienes permisos";
+                        return response;
+                    }
                     await DeleteAsync(model);
-                    response.Result = true;
+                    response.Result = ResponseStatus.Ok;
                     response.Message = "El usuario se ha borrado correctamente";
                 }
                 catch (Exception ex)
                 {
-                    response.Result = false;
+                    response.Result = ResponseStatus.InternalError;
                     response.Message = "Ha ocurrido un error al borrar el usuario";
                     Console.WriteLine($"Error: {ex.Message}");
                 }
